fix: api 내 에러처리 제거, api 코드 간소화, Nullable타입활용
