+++ src/apis/fetchTransferItems.ts
@@ -6,19 +6,6 @@ import { END_POINTS } from "@/constants/api";
 export const fetchTransferItems = async (): Promise<
   IReservation[] | undefined
 > => {
-  try {
-    const token = localStorage.getItem("accessToken");
-    if (!token) {
-      throw new Error("No Token");
-    }
-    const response = await axiosInstance.get(END_POINTS.RESERVATION, {
-      headers: {
-        Authorization: `${token}`,
-        "Content-Type": "application/json",
-      },
-    });
-    return response.data.data as IReservation[];
-  } catch (err) {
-    console.log(err, "err");
-  }
+  const { data } = await axiosInstance.get(END_POINTS.RESERVATION);
+  return data.data as IReservation[];
 };
