+++ src/components/carousel/Carousel.style.ts
@@ -3,10 +3,11 @@ import styled, { css } from "styled-components";
 
 export const CarouselContainer = styled.div<{
   $height: number;
+  $width: number;
 }>`
   position: relative;
 
-  width: 100%;
+  width: ${(props) => `${props.$width}px`};
   min-height: ${(props) => `${props.$height}px`};
   height: ${(props) => `${props.$height}px`};
