+++ src/components/carousel/Carousel.tsx
@@ -4,6 +4,7 @@ import * as S from "./Carousel.style.ts";
 
 interface CarouselProps {
   images: string[];
+  width?: number;
   height?: number;
   arrows?: boolean;
   infinite?: boolean;
@@ -13,6 +14,7 @@ interface CarouselProps {
 
 const Carousel = ({
   height = 300,
+  width = 300,
   images,
   arrows = true,
   infinite = false,
@@ -39,7 +41,7 @@ const Carousel = ({
   });
 
   return (
-    <S.CarouselContainer $height={height}>
+    <S.CarouselContainer $height={height} $width={width}>
       <S.SliderWrapper>
         <S.SliderContainer
           ref={sliderRef}
