+++ src/types/nullable.ts
@@ -1,0 +1,1 @@
+export type Nullable<T> = T | null;
